buddy/requirements.txt
fastapi
uvicorn[standard]
websockets
elevenlabs
openai
sounddevice
numpy
python-dotenv